Inclusion filtros customizados

+++ CrudLosCines/src/main/java/com/recursosformacion/lcs/service/EntradaService.java
@@ -17,8 +17,12 @@ import com.recursosformacion.lcs.util.Rutinas;
 @Service
 public class EntradaService implements IServicio<Entrada, Long> {
 
-	@Autowired
-	private IEntrada entradaRepository;
+	
+	private final IEntrada entradaRepository;
+	
+	EntradaService(IEntrada entradaRepository){
+		this.entradaRepository = entradaRepository;
+	}
 
 	@Override
 	public Entrada insert(Entrada entrada) {
@@ -38,11 +42,13 @@ public class EntradaService implements IServicio<Entrada, Long> {
 
 		Optional<Entrada> entradaDBO = entradaRepository.findById(entrada.getId_entrada());
 		if (entradaDBO.isEmpty()) {
-			throw new DAOException("El registro ya no existe");
+			throw new DAOException("El registro:" + entrada.getId_entrada() + ", ya no existe");
 		}
 		Entrada entradaDB = entradaDBO.get();
 
 		entradaDB.setIdCliente(Rutinas.nuevoSiNoVacio(entradaDB.getIdCliente(), entrada.getIdCliente()));
+		entradaDB.setEnt_fecha(Rutinas.nuevoSiNoVacio(entradaDB.getEnt_fecha(), entrada.getEnt_fecha()));
+		entradaDB.setEnt_fila(Rutinas.nuevoSiNoVacio(entradaDB.getEnt_fila(), entrada.getEnt_fila()));
 		entradaDB.setEnt_numero(Rutinas.nuevoSiNoVacio(entradaDB.getEnt_numero(), entrada.getEnt_numero()));
 		entradaDB.setEnt_cine(Rutinas.nuevoSiNoVacio(entradaDB.getEnt_cine(), entrada.getEnt_cine()));
 
@@ -50,7 +56,11 @@ public class EntradaService implements IServicio<Entrada, Long> {
 	}
 
 	@Override
-	public boolean deleteById(Long id_entrada) {
+	public boolean deleteById(Long id_entrada) throws DAOException {
+		Optional<Entrada> entradaDBO = entradaRepository.findById(id_entrada);
+		if (entradaDBO.isEmpty()) {
+			throw new DAOException("El registro:" + id_entrada + ", ya no existe");
+		}
 		 entradaRepository.deleteById(id_entrada);
 		 return true;
